+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - don't throw an error when `bootstrap()` is called multiple times w/o a callback
 
+### Changed
+- lint source before running build
+
 ## 1.0.0
 
 ### Changed